Adding tel: URL scheme for phone numbers in HTML
We can use the tel: URL scheme for phone numbers in HTML. Just like the mailto: URL scheme will open the default mail application. If the HTML page is viewed on a mobile phone and we select a link with the tel: scheme we can immediately call the number following the scheme. On a desktop computer, a VOIP call will be initiated.